Refactor transform gui into a single module
